finish version a-0.0.1
