+++ README.md
@@ -1,2 +1,41 @@
-# nintiacalc
-Daily reward and ROI calculator for Nintia Estate
+# Nintia Estate Calculator
+
+The Nintia Estate Calculator is a simple calculator for Nintia Estate players.
+
+The calculator allows players to calculate their possible daily profit according to their own personal investments.
+
+Read about how the calculations are made [HERE](./docu.md)
+
+## Preview
+
+![Preview1](./img/readme/prev1.png)
+![Preview2](./img/readme/prev2.png)
+
+## Live Demo
+
+[LIVE DEMO](https://orozcoding.github.io/tokenstoclaim/)
+
+## Built With
+
+- HTML, CSS and JavaScript
+
+## Authors
+
+👤 **Angel Orozco (orozCoding)**
+
+- GitHub: [@githubhandle](https://github.com/orozCoding)
+- LinkedIn: [LinkedIn](https://www.linkedin.com/in/angel-orozco-652230228/)
+
+## 🤝 Contributing
+
+Contributions, issues, and feature requests are welcome!
+
+Feel free to check the [issues page](../../issues/).
+
+## Show your support
+
+Give a ⭐️ if you like this project!
+
+## 📝 License
+
+This project is [MIT](./MIT.md) licensed.
