+++ docu.md
@@ -1,0 +1,124 @@
+# Nintia Calculator
+
+Nintia Estate Calculator es una calculadora basada en la información proporcionada en el [whitepaper](https://whitepaper.nintiaestate.com/v/espanol/) de [Nintia Estate](https://nintiaestate.com/).
+
+Todos los cálculos asumen un debido mantenimiento de las propiedades, y el uso más eficiente de los objetos. Por ejemplo, se asume que cada limpieza se está haciendo correctamente para obtener 50 NE por cada una, y no menos. Visita el [whitepaper](https://whitepaper.nintiaestate.com/v/espanol/) si necesitas más información al respecto.
+
+La intención inicialmente es calcular las ganancias y gastos fijos y de rutina, por lo que estos cálculos no incluyen cualquier operación especial o esporádica como la construcción o movimiento de una propieda, entre otras.
+
+## CÁLCULOS
+
+### RENTA
+
+Cada habitación produce 150 NE al día, de acuerdo al apartado [RENTA](https://whitepaper.nintiaestate.com/v/espanol/farming/rent) en el whitepaper.
+
+### TAREAS DIARIAS
+
+Visita la [TIENDA](https://game.nintiaestate.com/game/store) para corroborar los precios y usos que tomamos en cuenta para los cálculos.
+
+- LIMPIEZA
+
+Añadimos 40 NE por cada limpieza.
+
+15 limpiezas al día = 600 NE.
+
+Explicación:
+
+Cada limpieza a propiedades de otro usuario, y, usadas correctamnente, generan 50 NE por uso.
+
+Un kit de limpieza trae 50 usos.
+
+50 x 50 NE = 2.500 NE
+
+El costo del Kit es de 500 NE, por lo que el profit real es de:
+2.500 NE - 500 NE = 2.000 NE.
+
+Al dividir 2.000 entre 50 usos:
+
+2.000 ÷ 50 = 40 NE
+
+De esta forma, calculamos las ganancias diarias por limpieza, sacando de la ecuación de una vez el costo de inversión.
+
+- REPARACIÓN
+
+Añadimos 80 NE por cada reparación
+
+5 x 80 NE = 400 NE.
+
+Explicación:
+
+Cada reparación a propiedades de otro usuario generan 100 NE.
+
+Un kit de reparación trae 10 usos.
+
+10 x 100 NE = 1000 NE.
+
+El costo del Kit es de 200 NE, por lo que el profit real es de:
+1000 NE - 200 NE = 800 NE.
+
+Al dividir 800 entre 10 usos:
+
+800 ÷ 10 = 80 NE
+
+De esta forma, calculamos las ganancias diarias por reparación, sacando de la ecuación de una vez el costo de inversión.
+
+
+### GASTOS DIARIOS
+
+Otros gastos reptitivos pueden ser:
+
+- VALLAS
+
+Restamos 30 NE por cada valla (propiedad) al día.
+
+Explicación:
+
+Si tienes propiedades en el desierto, es posible que uses 1 valla al día, por cada propiedad.
+
+Una Valla tiene 10 usos y un costo de 300 NE.
+
+300 NE ÷ 10 = 30 NE.
+
+- SEGURO DE LIMPIEZA
+
+Restamos 80 NE por cada seguro de limpieza (propiedad) al día.
+
+Explicación:
+
+Si tienes propiedades en el desierto, es posible que uses 1 seguro de limpieza al día, por cada propiedad.
+
+Un seguro de limpieza tiene 5 usos y un costo de 400 NE.
+
+400 NE ÷ 5 = 80 NE.
+
+
+### CÁLCULO DE PROFIT
+
+- NE AL DÍA
+
+Renta + Tareas - Gastos
+
+- $NINTI AL DÍA
+
+La tasa de conversión de NE a $NINTI es de 125 NE = 1 $NINTI
+
+$NINTI al día se calcula fácilmente:
+
+NE AL DÍA ÷ 125
+
+- $USD AL DÍA
+
+Actualmente la app necesita el precio del $NINTI manualmente escrito por el usuario. En un futuro, automáticamente usando plataformas como CoinGecko.
+
+Para obtener los $USD al día, calculamos:
+
+$NINTI al día x Precio actual del $NINTI en $USD
+
+Una vez se tiene el precio del $NINTI en el mercado.
+
+## CRÉDITOS
+
+[Nintia Estate](https://nintiaestate.com/) es dueño de todos los diseños usados en el proyecto.
+
+Aplicación en desarrollo por [orozCoding](https://github.com/orozCoding), también conocido como [@ElProfeNFT](https://twitter.com/ElProfeNFT).
+
