+++ js/obj.js
@@ -5,6 +5,7 @@ const obj = {
   fence: 0,
   insurance: 0,
   price: 0,
+  ver: 'a-0.0.1'
 }
 
 export { obj };
